requirements.txt
cv2
ffmpeg
joblib
librosa
matplotlib
mediapipe
numpy
Pillow
scikit-learn
vit-pytorch